Use PostgreSQL LISTEN/NOTIFY to notify the scheduler that new jobs must to be executed. When an asynchronous job is submitted a trigger is fired that sends a notification. The scheduler looks for these notifications every 0.1 second, when a notification is found it forks a process to run the job. This avoid to poll the database at each loop of there is nothing new to execute.
The naptime configuration directive have been removed, the main loop
now waits 0.1 seconds between each loop. This is hardcoded and it
requires the installation of the Time::HiRes Perl modules.

For scheduled job the principle is the same except that the notification
invalidate the cache of scheduled jobs loaded at startup to inform the
scheduler that it must reload the jobs definitions from the database.