Execute the job code in quite the same way as Oracle by using the following DO block:
    DO $$
    DECLARE
        job bigint := $jobid;
        next_date timestamp with time zone := current_timestamp;
        broken boolean := false; -- Always false for asynchronous jobs
    BEGIN
        $WATH
    END;
    $$;

Not sure that the variable declaration could be useful but by this way
they can be used in the code to execute.

In Oracle the SQL that you submit in the what parameter is wrapped in
the following PL/SQL block:

    DECLARE
        job BINARY_INTEGER := :job;
        next_date DATE := :mydate;
        broken BOOLEAN := FALSE;
    BEGIN
        WHAT
        :mydate := next_date;
        IF broken THEN :b := 1; ELSE :b := 0; END IF;
    END;

This commit also refactor some code and change the data type of column
status in table dbms_job.all_scheduler_job_run_details to text instead
of varchar(128). With the DO block the error message are larger and
actually we don't need a size limitation here.