Asynchronous job processing now also look at scheduled jobs without an interval value. These jobs was submitted with an execution date in the future. To be able to process these jobs after their creation we can not rely on NOTIFY so we need to collect these jobs regulary to look if they should be executed and removed.
Add job_queue_interval to force the collect of the scheduled jobs with
no interval to execute them in case their next_date value is over. The
default is to search for these jobs every 5 seconds since the last
inactive loop (no notification was received). That mean that a job will
be executed at worst job_queue_interval seconds after is requested
execution date.  This value is hardcoded for the moment but could be a
GUC in the future.