Add support to scheduled jobs. Like for asynchronous jobs, the daemon on notification from the server (when a job is submitted or when a job is modified) or periodically at job_queue_interval it will look for jobs with a next_date <= current_timestamp and no in broken state. If such job are found it will update the next execution date from the command specified in the interval column and execute the job. Once it ends the job will be updated to set the start date of the execution (this_date) and the number of failures since last successful run if there was an error. Or it will update the date of the last successful run (last_date), set the execution time of the job (total_time) and set the failures counter to 0.
As the server send notification to the daemon when there is a
modification on table dbms_job.all_scheduled_jobs we use internally the
unused column instance to instruct the trigger to not call NOTIFY when
the value of the instance column change.

Once the information into dbms_job.all_scheduled_jobs have been modified
it log the job run details into dbms_job.all_scheduler_job_run_details.