Allow strftime() escapes in log filename, for example to have a log file per week day use %a in the file name.
Add new configuration directive log_truncate_on_rotation to truncate file
on rotation. When activated an existing log file with the same name as the
new log file will be truncated rather than appended to. But such truncation
only occurs on time-driven rotation, not on restarts.

Thanks to K RamaKrishna Sastry for the feature request.